# apidocs/docs/swagger.yaml
basePath: /api/v1
definitions:
  controller.Account:
    properties:
      id:
        example: 1
        format: int64
        type: integer
      name:
        example: account name
        type: string
    type: object
  controller.HTTPError:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample server celler server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /accounts:
    get:
      consumes:
      - application/json
      description: get accounts
      parameters:
      - description: name search by q
        format: email
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/controller.Account'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.HTTPError'
      summary: List accounts
      tags:
      - accounts
  /accounts/{id}:
    get:
      consumes:
      - application/json
      description: get string by ID
      parameters:
      - description: Account ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.Account'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/controller.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.HTTPError'
      summary: Show an account
      tags:
      - accounts
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"